+++ app/src/main/java/com/practice/hanbitlunch/MainActivity.kt
@@ -10,7 +10,7 @@ import androidx.compose.ui.Modifier
 import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
 import androidx.work.WorkManager
 import com.practice.designsystem.theme.BlindarTheme
-import com.practice.hanbitlunch.screen.navigation.BlindarNavHost
+import com.practice.hanbitlunch.screen.BlindarNavHost
 import com.practice.preferences.PreferencesRepository
 import com.practice.work.setOneTimeFetchMealWork
 import com.practice.work.setOneTimeFetchScheduleWork
